+++ src/environments/environment.preview.ts
@@ -3,4 +3,5 @@ export const environment = {
   clerkPublishableKey: 'pk_live_Y2xlcmsubG9uZG9uY2hlc3MuY2Ek',
   googleMapsApiKey: 'AIzaSyCeuTgq4qe2k8obnbYWrwWsG91B6aDUSA0',
   lccApiBaseUrl: 'https://api.londonchess.ca/v1',
+  sentryDsn: '',
 };